# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Q_ORDER API",
    "description": "Q_ORDER Backend API Documentation - Multi-Tenant Restaurant Ordering System",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token. Example: Bearer YOUR_TOKEN"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/kds/orders": {
      "get": {
        "summary": "Get orders for KDS",
        "description": "Retrieves orders that are relevant to the Kitchen Display System.",
        "tags": [
          "KDS"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/kds/items/{item_id}/status": {
      "patch": {
        "summary": "Update KDS item status",
        "description": "Updates the preparation/status state of an order item in the Kitchen Display System.",
        "tags": [
          "KDS"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order item ID"
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "New status of the order item",
                    "example": "preparing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item status updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/kds/sync": {
      "get": {
        "summary": "Synchronize KDS",
        "description": "Retrieves the latest data required to synchronize the Kitchen Display System.",
        "tags": [
          "KDS"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "KDS synchronized successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/kds/print-kot": {
      "post": {
        "summary": "Print KOT",
        "description": "Sends a Kitchen Order Ticket (KOT) for printing.",
        "tags": [
          "KDS"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "order_id": "123e4567-e89b-12d3-a456-426614174000"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "KOT sent for printing successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/orders": {
      "post": {
        "summary": "Submit a new order",
        "description": "Creates and submits a new customer order.",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Unique UUID v4 key used to prevent duplicate order submissions."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "session_id": "123e4567-e89b-12d3-a456-426614174000",
                "items": [
                  {
                    "menu_item_id": "123e4567-e89b-12d3-a456-426614174001",
                    "quantity": 2
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order submitted successfully"
          },
          "400": {
            "description": "Invalid order data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Duplicate order request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/orders/session-history": {
      "get": {
        "summary": "Get session order history",
        "description": "Retrieves the order history associated with the current customer session.",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Session order history retrieved successfully"
          },
          "400": {
            "description": "Invalid session information"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/status": {
      "get": {
        "summary": "Get order status",
        "description": "Retrieves the current status of a specific order.",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Order ID"
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Order status retrieved successfully"
          },
          "400": {
            "description": "Invalid order ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update order status",
        "description": "Updates the status of a specific order.",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Order ID"
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tenant ID"
          },
          {
            "in": "header",
            "name": "X-Branch-ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "New order status",
                    "example": "preparing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully"
          },
          "400": {
            "description": "Invalid order status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}